pyproject.toml
[tool.poetry]
name = "integrations"
version = "0.1.0"
description = ""
authors = ["Neil Munro <neil@elixirsoftware.co.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
black = "^23.1.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
coverage = "^7.2.1"
ipython = "^8.11.0"
sphinx = "^6.1.3"
django-polymorphic = "^3.1.0"
djangorestframework = "^3.14.0"
markdown = "^3.4.1"
django-filter = "^22.1"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
